QuickFit3/fcs_maxent:   - fixed bug in datacut sliders   - added SPinBox to choose iterations/click   - fixed some more minor bugs   - implemented Reset buttons   - fixed bug in display of evaluation results (table in QFEvaluationPropertyEditor)
QuickFit3/stage_pi: fixed a signed/unsigned comparison warning
QuickFit3/cam_systemcamera: fixed a signed/unsigned comparison warning
QuickFit3/filterc_test: fixed a signed/unsigned comparison warning
QuickFit3/picoquantimporters: fixed a signed/unsigned comparison warning

git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@1405 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/cam_systemcamera/qfecamsystemcamera.cpp
@@ -92,7 +92,7 @@ void QFECamSystemcamera::initExtension() {
     cameras=cap.getDeviceList();
     QMenu* scm=new QMenu(tr("&System Camera Controls"));
     scm->setIcon(QIcon(":/cam_systemcam/qfecamsystemcamera.png"));
-    for (int i=0; i<cameras.size(); i++) {
+    for (size_t i=0; i<cameras.size(); i++) {
         vids.append(new VideoCapture());
         if (vids[i]->getConfigDialogCount()>0) {
             QMenu* m=scm->addMenu(cameras[i].c_str());
